+++ pkg/cvo/cvo.go
@@ -938,7 +938,7 @@ func hasReachedLevel(cv *configv1.ClusterVersion, desired configv1.Update) bool 
 
 func (optr *Operator) defaultPreconditionChecks() precondition.List {
 	return []precondition.Precondition{
-		preconditioncv.NewRollback(optr.currentVersion),
+		preconditioncv.NewMetadataUpdate(optr.currentVersion),
 		preconditioncv.NewUpgradeable(optr.cvLister),
 		preconditioncv.NewRecentEtcdBackup(optr.cvLister, optr.coLister),
 		preconditioncv.NewRecommendedUpdate(optr.cvLister),
